Staging

See merge request ocelotbot/OcelotBOTv5!190

+++ modules/support.js
@@ -101,7 +101,7 @@ module.exports = class SupportServer {
             try {
                 this.bot.logger.warn(`Checking domain in ${message.guild.id} (${message.content})`);
                 let result = await axios.post("https://anti-fish.bitflow.dev/check", {message: message.content}).catch(()=>null);
-                if (result?.data?.match) {
+                if (result?.data?.match && result.data.matches[0]?.trust_rating > 0.5) { // Hacky
                     this.bot.logger.warn(`Deleting possible free nitro message ${message.content}`);
                     const isAdmin = message.member.permissions.has("ADMINISTRATOR");
                     if(!isAdmin)message.delete();
